Wrapper function around hydroGOF functions

Calculates Goodness-of-Fit functions for two runoff series

Usage

calc_hydroGOF(GOF_fun, Qsim, Qobs, na.rm = TRUE)

Arguments

GOF_fun

a function, or (todo consider only a functional) a string with function name, of the format GOF_fun(Qsim, Qobs, na.rm = "TRUE"), typically from the hydroGOF package

Qsim

vector, matrix, data.frame etc of simulated runoff values

Qobs

vector, matrix, data.frame etc of observed runoff values

na.rm

a logical value indicating if NA should be removed

Value

transformed runoff in same format as input

Note

This function requires the hydroGOF package to be installed. Not imported as this package depends on deprecated packages as sp etc.

See also

Examples

# hydroGOF must be loaded
library(hydroGOF)
#> Loading required package: zoo
#> 
#> Attaching package: 'zoo'
#> The following objects are masked from 'package:base':
#> 
#>     as.Date, as.Date.numeric
#> The legacy packages maptools, rgdal, and rgeos, underpinning the sp package,
#> which was just loaded, will retire in October 2023.
#> Please refer to R-spatial evolution reports for details, especially
#> https://r-spatial.org/r/2023/05/15/evolution4.html.
#> It may be desirable to make the sf package available;
#> package maintainers should consider adding sf to Suggests:.
#> The sp package is now running under evolution status 2
#>      (status 2 uses the sf package in place of rgdal)
#> Please note that 'maptools' will be retired during October 2023,
#> plan transition at your earliest convenience (see
#> https://r-spatial.org/r/2023/05/15/evolution4.html and earlier blogs
#> for guidance);some functionality will be moved to 'sp'.
#>  Checking rgeos availability: FALSE
calc_hydroGOF(KGE, 1:10, seq(0, 9))
#> [1] 0.7777778
# this is NA
calc_hydroGOF("KGE", 1:10, rep(0, 10))
#> Warning: the standard deviation is zero
#> [1] NA
# this is also NA
calc_hydroGOF(KGE, 1:10, as.numeric(rep(NA, 10)))
#> Warning: 'sim' and 'obs' are empty or they do not have any common pair of elements with data !!
#> Warning: There are no pairs of 'sim' and 'obs' without missing values !
#> [1] NA