Create Default Calibration Settings for airGR Models

Generates a list of default calibration settings for a specified model from the airGR family. This includes parameter bounds and optimizer configurations for DEoptim, malschains, and hydroPSO.

Usage

set_airGR_par(model)

Arguments

model

A string specifying the airGR model (e.g., "GR4J", "CemaNeigeGR4J").

Value

A list containing:

lower, upper – numeric vectors of parameter bounds.

nof_param – number of parameters.

DEoptim, malschains, hydroPSO – optimizer settings.

has_snow_module – logical indicating whether the model includes a snow module.

Details

The function uses CreateCalibOptions to extract parameter bounds and sets default configurations for several optimizers. It also detects whether the model includes a snow module based on the model name.

Note

This function only supports models from the airGR family. An error is thrown otherwise.

Examples

set_airGR_par("GR4J")
#> $lower
#> [1]  4.585621e-05 -1.090365e+04  4.585621e-05  5.000000e-01
#> 
#> $upper
#> [1] 21807.30 10903.65 21807.30    20.00
#> 
#> $nof_param
#> [1] 4
#> 
#> $DEoptim
#> $DEoptim$NP
#> [1] 40
#> 
#> $DEoptim$itermax
#> [1] 200
#> 
#> 
#> $malschains
#> $malschains$maxEvals
#> [1] 2000
#> 
#> 
#> $hydroPSO
#> $hydroPSO$control
#> $hydroPSO$control$write2disk
#> [1] FALSE
#> 
#> $hydroPSO$control$verbose
#> [1] FALSE
#> 
#> $hydroPSO$control$npart
#> [1] 80
#> 
#> $hydroPSO$control$maxit
#> [1] 50
#> 
#> $hydroPSO$control$reltol
#> [1] 1e-10
#> 
#> 
#> 
#> $has_snow_module
#> [1] FALSE
#> 
set_airGR_par("CemaNeigeGR4J")
#> $lower
#> [1]  4.585621e-05 -1.090365e+04  4.585621e-05  5.000000e-01  0.000000e+00
#> [6]  2.292810e-07
#> 
#> $upper
#> [1] 21807.2988 10903.6494 21807.2988    20.0000     1.0000   109.0365
#> 
#> $nof_param
#> [1] 6
#> 
#> $DEoptim
#> $DEoptim$NP
#> [1] 60
#> 
#> $DEoptim$itermax
#> [1] 200
#> 
#> 
#> $malschains
#> $malschains$maxEvals
#> [1] 2000
#> 
#> 
#> $hydroPSO
#> $hydroPSO$control
#> $hydroPSO$control$write2disk
#> [1] FALSE
#> 
#> $hydroPSO$control$verbose
#> [1] FALSE
#> 
#> $hydroPSO$control$npart
#> [1] 80
#> 
#> $hydroPSO$control$maxit
#> [1] 50
#> 
#> $hydroPSO$control$reltol
#> [1] 1e-10
#> 
#> 
#> 
#> $has_snow_module
#> [1] TRUE
#>