Validate model

Calculates validation measures for different transformation types

Usage

validate_model(Qsim, Qobs, val_crit_transfo = "KGE__none")

Arguments

Qsim

vector with simulated runoff

Qobs

vector with observed runoff

val_crit_transfo

a vector of strings specifying validation criteria and a runoff transformation separated by a "__". Supported are validation criteria from the hydroGOF package usable by the calc_hydroGOF function, for supported runoff transformations please refer to transfo_q

Value

a long data frame with columns crit indicating the used validation criterion, transfo for the used runoff transformation and value.

Examples

validate_model(
  1:10, seq(2, 11),
  c("KGE__log", "NSE__inv", "VE__none", "pbias__none")
)
#> # A tibble: 4 x 4
#>   crit  transfo lambda   value
#>   <chr> <chr>   <chr>    <dbl>
#> 1 KGE   log     NA       0.650
#> 2 NSE   inv     NA      -0.930
#> 3 VE    none    NA       0.846
#> 4 pbias none    NA     -15.4